requirements.txt
django
djangorestframework
pyyaml
redis
celery
psycopg2-binary
python-dotenv
docker
djangorestframework-simplejwt
drf-spectacular
channels
channels-Redis
